improved demo
