+++ package.json
@@ -15,7 +15,8 @@
     "rollup-plugin-svelte": "~6.1.1",
     "rollup-plugin-terser": "^5.1.2",
     "sirv-cli": "^0.4.4",
-    "svelte": "^3.24.0"
+    "svelte": "^3.24.0",
+    "svelte-highlight": "^3.0.0"
   },
   "scripts": {
     "build": "rollup -c",
